[#60]rename: Util 클래스를 하나의 파일단위로 관리하도록 수정
