+++ app/visualize/analysis/stmt/parser/expr/parser/binop_expr.py
@@ -1,7 +1,7 @@
 import ast
 
 from app.visualize.analysis.stmt.parser.expr.models.expr_obj import ExprObj, BinopObj
-from app.visualize.utils.util import Util
+from app.visualize.utils.utils import Util
 
 
 class BinopExpr:
