+++ app/visualize/analysis/stmt/parser/expr/parser/list_expr.py
@@ -1,5 +1,5 @@
 from app.visualize.analysis.stmt.parser.expr.models.expr_obj import ListObj, ExprObj
-from app.visualize.utils.util import Util
+from app.visualize.utils.utils import Util
 
 
 class ListExpr:
